Move math utils to martin-tile-utils
Mathematics should be consolidated in the utils crate. Eventually, I hope we will move it to `tile-grid` or `utile` crates and consolidate our efforts there.
