+++ mbtiles/src/summary.rs
@@ -8,7 +8,7 @@ use std::fmt::{Display, Formatter};
 use std::path::PathBuf;
 use std::str::FromStr;
 
-use martin_tile_utils::{EARTH_CIRCUMFERENCE, EARTH_RADIUS};
+use martin_tile_utils::{get_zoom_precision, xyz_to_bbox};
 use serde::Serialize;
 use size_format::SizeFormatterBinary;
 use sqlx::{query, SqliteExecutor};
@@ -158,7 +158,8 @@ impl Mbtiles {
                         r.min_tile_y.unwrap(),
                         r.max_tile_x.unwrap(),
                         r.max_tile_y.unwrap(),
-                    ),
+                    )
+                    .into(),
                 }
             })
             .collect();
@@ -186,65 +187,13 @@ impl Mbtiles {
     }
 }
 
-/// Convert min/max XYZ tile coordinates to a bounding box
-fn xyz_to_bbox(zoom: u8, min_x: i32, min_y: i32, max_x: i32, max_y: i32) -> Bounds {
-    let tile_size = EARTH_CIRCUMFERENCE / f64::from(1_u32 << zoom);
-    let (min_lng, min_lat) = webmercator_to_wgs84(
-        -0.5 * EARTH_CIRCUMFERENCE + f64::from(min_x) * tile_size,
-        -0.5 * EARTH_CIRCUMFERENCE + f64::from(min_y) * tile_size,
-    );
-    let (max_lng, max_lat) = webmercator_to_wgs84(
-        -0.5 * EARTH_CIRCUMFERENCE + f64::from(max_x + 1) * tile_size,
-        -0.5 * EARTH_CIRCUMFERENCE + f64::from(max_y + 1) * tile_size,
-    );
-
-    Bounds::new(min_lng, min_lat, max_lng, max_lat)
-}
-
-fn get_zoom_precision(zoom: u8) -> usize {
-    let lng_delta = webmercator_to_wgs84(EARTH_CIRCUMFERENCE / f64::from(1_u32 << zoom), 0.0).0;
-    let log = lng_delta.log10() - 0.5;
-    if log > 0.0 {
-        0
-    } else {
-        -log.ceil() as usize
-    }
-}
-
-fn webmercator_to_wgs84(x: f64, y: f64) -> (f64, f64) {
-    let lng = (x / EARTH_RADIUS).to_degrees();
-    let lat = (f64::atan(f64::sinh(y / EARTH_RADIUS))).to_degrees();
-    (lng, lat)
-}
-
 #[cfg(test)]
 mod tests {
     #![allow(clippy::unreadable_literal)]
 
-    use approx::assert_relative_eq;
     use insta::assert_yaml_snapshot;
 
-    use crate::summary::webmercator_to_wgs84;
     use crate::{init_mbtiles_schema, MbtResult, MbtType, Mbtiles};
-
-    #[actix_rt::test]
-    async fn meter_to_lng_lat() {
-        let (lng, lat) = webmercator_to_wgs84(-20037508.34, -20037508.34);
-        assert_relative_eq!(lng, -179.99999991016847, epsilon = f64::EPSILON);
-        assert_relative_eq!(lat, -85.05112877205713, epsilon = f64::EPSILON);
-
-        let (lng, lat) = webmercator_to_wgs84(20037508.34, 20037508.34);
-        assert_relative_eq!(lng, 179.99999991016847, epsilon = f64::EPSILON);
-        assert_relative_eq!(lat, 85.05112877205713, epsilon = f64::EPSILON);
-
-        let (lng, lat) = webmercator_to_wgs84(0.0, 0.0);
-        assert_relative_eq!(lng, 0.0, epsilon = f64::EPSILON);
-        assert_relative_eq!(lat, 0.0, epsilon = f64::EPSILON);
-
-        let (lng, lat) = webmercator_to_wgs84(3000.0, 9000.0);
-        assert_relative_eq!(lng, 0.02694945851388753, epsilon = f64::EPSILON);
-        assert_relative_eq!(lat, 0.0808483487118794, epsilon = f64::EPSILON);
-    }
 
     #[actix_rt::test]
     async fn summary_empty_file() -> MbtResult<()> {
@@ -356,7 +305,7 @@ mod tests {
               - -123.75000000000001
               - -40.97989806962013
               - 180
-              - 61.60639637138627
+              - 61.60639637138628
           - zoom: 6
             tile_count: 72
             min_tile_size: 64
@@ -366,7 +315,7 @@ mod tests {
               - -123.75000000000001
               - -40.97989806962013
               - 180
-              - 61.60639637138627
+              - 61.60639637138628
         "###);
 
         Ok(())
