+++ tests/expected/martin-cp/flat-with-hash_summary.txt
@@ -7,7 +7,7 @@ Page size: 512B
     1 |         4 |      474B |      983B |      609B | -180,-85,180,85
     2 |         5 |      150B |      865B |      451B | -90,-67,180,67
     3 |         8 |       57B |      839B |      264B | -45,-41,180,67
-    4 |        13 |       57B |      751B |      216B | -22,-22,157,56
+    4 |        13 |       57B |      751B |      216B | -23,-22,158,56
     5 |        27 |       57B |      666B |      167B | -11,-11,146,49
     6 |        69 |       57B |      636B |      127B | -6,-6,146,45
   all |       127 |       57B |      983B |      187B | -180,-85,180,85
