+++ tests/expected/martin-cp/flat_summary.txt
@@ -7,7 +7,7 @@ Page size: 512B
     1 |         2 |      150B |      172B |      161B | -180,-85,0,85
     2 |         4 |      291B |      690B |      414B | -90,-67,90,67
     3 |         7 |       75B |      727B |      263B | -45,-41,90,67
-    4 |        13 |       75B |      684B |      225B | -22,-22,157,56
+    4 |        13 |       75B |      684B |      225B | -23,-22,158,56
     5 |        27 |       75B |      659B |      195B | -11,-11,146,49
     6 |        69 |       75B |      633B |      155B | -6,-6,146,45
   all |       123 |       75B |      727B |      190B | -180,-85,180,85
